requirements.txt
# FastAPI 백엔드 서버 필수 패키지
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3

# 데이터베이스 ORM 및 인증
sqlalchemy==2.0.35
alembic==1.14.0  # 데이터베이스 마이그레이션
sqlite3  # 기본 내장 (개발용)

# 인증 및 보안
python-jose[cryptography]==3.3.0  # JWT 토큰
passlib[bcrypt]==1.7.4  # 패스워드 해싱
python-multipart==0.0.12  # OAuth2 폼 데이터

# OAuth (카카오 로그인)
authlib==1.3.2  # OAuth2 클라이언트
httpx==0.28.1  # 비동기 HTTP 클라이언트

# 스케줄러 (자동 업데이트용)
apscheduler==3.11.0

# 기존 lott.py 의존성
tqdm==4.67.1

# 추가 유틸리티
requests==2.32.3  # HTTP 요청 (옵션)

# 개발 및 배포 (옵션)
python-dotenv==1.0.1  # 환경변수 관리
